refactor: Seperate mcp servers and integrations types
