+++ portkey_ai/api_resources/apis/mcp_servers.py
@@ -3,7 +3,7 @@ from portkey_ai._vendor.openai import NOT_GIVEN, NotGiven
 from portkey_ai.api_resources.base_client import APIClient, AsyncAPIClient
 from urllib.parse import urlencode
 from portkey_ai.api_resources.apis.api_resource import APIResource, AsyncAPIResource
-from portkey_ai.api_resources.types.mcp_type import (
+from portkey_ai.api_resources.types.mcp_servers_type import (
     McpServerCreateResponse,
     McpServerRetrieveResponse,
     McpServerListResponse,
